first test

+++ protected/controllers/ProfileController.php
@@ -23,6 +23,11 @@ class ProfileController extends Controller
 				#'ajax'=>false,
 				'users'=>array('@'),
 			),
+			array('allow',  // just guest can perform 'activate' and 'login' actions
+				'actions'=>array('MyareaJsonView'),
+				'users'=>array('*'),
+			),
+			
 			array('deny',  // deny all users
 				'users'=>array('*'),
 			),
@@ -117,46 +122,65 @@ class ProfileController extends Controller
 	public function actionMyarea()
 	{
 		$cs=Yii::app()->getClientScript();
-        $cs->registerCssFile(Yii::app()->request->baseUrl.'/css/add_form.css');
+        $cs->registerCssFile('/css/add_form.css');
         $cs->registerScriptFile('http://api-maps.yandex.ru/1.1/index.xml?key='.$this->mapkey);
         $model=$this->loadModel(Yii::app()->user->id);	
         
-        	if(isset($_POST['UserAreaShapes']))
+        	if(isset($_POST['UserAreaShapes']) || isset($_POST['UserGroupsUser']))
 			{
 				$ids=Array();
-				foreach ($_POST['UserAreaShapes'] as $i=>$shape)
-					{
-						if ($shape['id']) $ids[]=$shape['id'];	
-					}
-				if ($ids){
-				$delshapes=UserAreaShapes::model()->findAll('id NOT IN ('.implode(',',$ids).') AND ug_id='.$model->id);				
-				foreach ($delshapes as $shape) $shape->delete();
-				}
-				
-				foreach ($_POST['UserAreaShapes'] as $i=>$shape)
-					{
-						$shapemodel=$shape['id'] ? UserAreaShapes::model()->findByPk((int)$shape['id']) : new UserAreaShapes;
-						if ($shapemodel){
-							$shapemodel->ug_id=$model->id;
-							$shapemodel->ordering=$shape['ordering'];
-							if ($shapemodel->isNewRecord) $shapemodel->save();
-							$ii=0;
-							foreach ($_POST['UserAreaShapePoints'][$i] as $point)
-							{
-								if ($ii>=$shapemodel->countPoints) break;
-								$pointmodel=$point['id'] ? UserAreaShapePoints::model()->findByPk((int)$point['id']) : new UserAreaShapePoints;				
-								$pointmodel->attributes=$point;
-								$pointmodel->shape_id=$shapemodel->id;
-								$pointmodel->save();
-								$ii++;
-							}
-							if (!$shapemodel->points) $shapemodel->delete();
+				if(isset($_POST['UserAreaShapes'])){
+					foreach ($_POST['UserAreaShapes'] as $i=>$shape)
+						{
+							if (isset($shape['id']) && $shape['id']) $ids[]=$shape['id'];	
 						}
-					}	
+				}
+				if ($ids) UserAreaShapes::model()->deleteAll('id NOT IN ('.implode(',',$ids).') AND ug_id='.$model->id);				
+				else UserAreaShapes::model()->deleteAll('ug_id='.$model->id);
+				
+				if(isset($_POST['UserAreaShapes'])){
+					foreach ($_POST['UserAreaShapes'] as $i=>$shape)
+						{
+							$shapemodel=$shape['id'] ? UserAreaShapes::model()->findByPk((int)$shape['id']) : new UserAreaShapes;
+							if ($shapemodel){
+								$shapemodel->ug_id=$model->id;
+								$shapemodel->ordering=$shape['ordering'];
+								if ($shapemodel->isNewRecord) $shapemodel->save();
+								$ii=0;
+								foreach ($_POST['UserAreaShapePoints'][$i] as $point)
+								{
+									if ($ii>=$shapemodel->countPoints) break;
+									$pointmodel=$point['id'] ? UserAreaShapePoints::model()->findByPk((int)$point['id']) : new UserAreaShapePoints;				
+									$pointmodel->attributes=$point;
+									$pointmodel->shape_id=$shapemodel->id;
+									$pointmodel->save();
+									$ii++;
+								}
+								if (!$shapemodel->points) $shapemodel->delete();
+							}
+						}							
 				$this->redirect(array('/holes/myarea'));
+				}
 			}        
       	$this->render('myarea',array('model'=>$model));
 	}
+	
+	
+	public function actionMyareaJsonView($user_id=0)
+	{
+		if (!$user_id) $user_id=Yii::app()->user->id;
+        $model=$this->loadModel((int)$user_id);
+        $pointsArr=Array();
+        foreach ($model->hole_area as $ind=>$shape) {
+				foreach ($shape->points as $i=>$point) {
+					$pointsArr[$ind][]=Array('lat'=>$point->lat, 'lng'=>$point->lng);
+				} 
+			}
+		echo $_GET['jsoncallback'].'({"area": '.CJSON::encode($pointsArr).'})';
+		
+		Yii::app()->end();		
+
+	}	
 	
 	public function actionMyareaAddshape()
 	{
